src/main/java/org/entando/kubernetes/controller/digitalexchange/component/PostInitResourceController.java
/*
 * Copyright 2018-Present Entando Inc. (http://www.entando.com) All rights reserved.
 *
 * This library is free software; you can redistribute it and/or modify it under
 * the terms of the GNU Lesser General Public License as published by the Free
 * Software Foundation; either version 2.1 of the License, or (at your option)
 * any later version.
 *
 * This library is distributed in the hope that it will be useful, but WITHOUT
 * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
 * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
 * details.
 */

package org.entando.kubernetes.controller.digitalexchange.component;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.entando.kubernetes.service.digitalexchange.job.PostInitStatusService;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;

@Slf4j
@RestController
@RequiredArgsConstructor
public class PostInitResourceController implements PostInitResource {

    private final PostInitStatusService status;
    private static final String ERROR_MSG = "Can't manage a request with bundleCode: ";

    @Override
    public ResponseEntity<String> getStatus() {
        return ResponseEntity.ok(status.getStatus().toString());
    }
}